Setting up ...

+++ src/laplacian.c
@@ -40,7 +40,7 @@ void ax_setup(struct gs_data **gsh, double **weights, struct comm *c, \
   rsb_setup = 1;
 }
 //------------------------------------------------------------------------------
-void ax(Vector *v, Vector *u, struct gs_data *gsh, double *weights, int64 nc) {
+void ax(Vector *v, Vector *u, struct gs_data *gsh, double *weights, int32 nc) {
   if (rsb_setup == 0) {
     fprintf(stderr, "Need to call ax_setup before this routine.");
     return;
